requirements.txt
Django==4.1.1
psycopg2-binary==2.9.3
invoke==1.7.1
uwsgi==2.0.20
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
social-auth-app-django==5.0.0
celery==5.2.7
redis==4.3.4
